moving helper functions to tools.py
